+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- Add ability to expand/collapse all groups in table (roubkar)
+
 ## 1.1.11 Nov 9 2020
 - Display group config in a group header (roubkar, gorarakelyan)
 - Sort runs by date in Dashboard screen (roubkar)